Remove FsTokenPersistence
Closes #5665
